+++ db/updates/update_9.8-9.9.sql
@@ -1,2 +1,3 @@
 ALTER TABLE `makeFormula` ADD `notes` MEDIUMTEXT NULL AFTER `originalQuantity`; 
 ALTER TABLE `makeFormula` ADD `skip` INT NOT NULL DEFAULT '0' AFTER `notes`; 
+ALTER TABLE `makeFormula` ADD `replacement_id` INT NOT NULL DEFAULT '0' AFTER `ingredient_id`;